+++ packages/block-editor/src/index.js
@@ -7,6 +7,7 @@ import '@wordpress/rich-text';
  * Internal dependencies
  */
 import './hooks';
+export { getInlineStyles as __experimentalGetInlineStyles } from './hooks';
 export * from './components';
 export * from './utils';
 export { storeConfig, store } from './store';